Replace Ollama LLM integration with LMStudios integration while maintaining all existing functionality (sentiment tracking, chat handling, event system, etc.)
