+++ LMStudioChat.cmake
@@ -21,14 +21,14 @@ if(TARGET modules)
     if(OpenSSL_FOUND OR OPENSSL_FOUND)
         target_compile_definitions(modules PRIVATE CPPHTTPLIB_OPENSSL_SUPPORT)
         target_link_libraries(modules PRIVATE OpenSSL::SSL OpenSSL::Crypto)
-        message(STATUS "[mod-ollama-chat] OpenSSL found - HTTPS support enabled")
+        message(STATUS "[LMStudioChat] OpenSSL found - HTTPS support enabled")
     else()
-        message(WARNING "[mod-ollama-chat] OpenSSL not found - only HTTP support available")
-        message(STATUS "[mod-ollama-chat] To enable HTTPS: install OpenSSL development libraries")
+        message(WARNING "[LMStudioChat] OpenSSL not found - only HTTP support available")
+        message(STATUS "[LMStudioChat] To enable HTTPS: install OpenSSL development libraries")
         if(WIN32)
-            message(STATUS "[mod-ollama-chat] Windows: Install vcpkg and run 'vcpkg install openssl'")
+            message(STATUS "[LMStudioChat] Windows: Install vcpkg and run 'vcpkg install openssl'")
         else()
-            message(STATUS "[mod-ollama-chat] Linux: apt install libssl-dev (Ubuntu/Debian) or yum install openssl-devel (RHEL/CentOS)")
+            message(STATUS "[LMStudioChat] Linux: apt install libssl-dev (Ubuntu/Debian) or yum install openssl-devel (RHEL/CentOS)")
         endif()
     endif()
