+++ data/sql/characters/base/2025_05_30_personalities.sql
@@ -1,8 +1,8 @@
 -- Drop the existing table if it exists
-DROP TABLE IF EXISTS `mod_ollama_chat_personality`;
+DROP TABLE IF EXISTS `LMStudioChat_personality`;
 
 -- Create the new table with BIGINT guid and VARCHAR personality
-CREATE TABLE IF NOT EXISTS `mod_ollama_chat_personality` (
+CREATE TABLE IF NOT EXISTS `LMStudioChat_personality` (
   `guid` BIGINT NOT NULL,
   `personality` VARCHAR(64) NOT NULL,
   PRIMARY KEY (`guid`)
