+++ data/sql/characters/base/2025_05_31_personality_template.sql
@@ -1,9 +1,9 @@
-CREATE TABLE IF NOT EXISTS `mod_ollama_chat_personality_templates` (
+CREATE TABLE IF NOT EXISTS `LMStudioChat_personality_templates` (
   `key` VARCHAR(64) NOT NULL PRIMARY KEY,
   `prompt` TEXT NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
 
-INSERT INTO `mod_ollama_chat_personality_templates` (`key`, `prompt`) VALUES
+INSERT INTO `LMStudioChat_personality_templates` (`key`, `prompt`) VALUES
 ('GAMER', 'Focus on game mechanics, min-maxing, and efficiency.'),
 ('ROLEPLAYER', 'Respond in-character, weaving lore into your response.'),
 ('LOOTGOBLIN', 'Talk about rare loot, gold strategies, and treasure hunting.'),
