+++ data/sql/characters/base/2025_06_14_chat_history.sql
@@ -1,5 +1,5 @@
 
-CREATE TABLE IF NOT EXISTS mod_ollama_chat_history (
+CREATE TABLE IF NOT EXISTS LMStudioChat_history (
     id BIGINT UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
     bot_guid BIGINT UNSIGNED NOT NULL,
     player_guid BIGINT UNSIGNED NOT NULL,
