+++ data/sql/characters/base/2025_07_24_sentiment_tracking.sql
@@ -1,4 +1,4 @@
-CREATE TABLE IF NOT EXISTS mod_ollama_chat_bot_player_sentiments (
+CREATE TABLE IF NOT EXISTS LMStudioChat_bot_player_sentiments (
     id BIGINT UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
     bot_guid BIGINT UNSIGNED NOT NULL,
     player_guid BIGINT UNSIGNED NOT NULL,
