+++ src/LMStudioChat_config.h
@@ -1,0 +1,223 @@
+
+
+
+#ifndef LMSTUDIO_CHAT_CONFIG_H
+#define LMSTUDIO_CHAT_CONFIG_H
+
+#include <string>
+#include <cstdint>
+#include <vector>
+#include <deque>
+#include <unordered_map>
+#include <mutex>
+#include <ctime>
+#include "ScriptMgr.h"  // Ensure WorldScript is defined
+
+// --------------------------------------------
+// Distance/Range Configuration
+// --------------------------------------------
+extern float      g_SayDistance;
+extern float      g_YellDistance;
+extern float      g_RandomChatterRealPlayerDistance;
+extern float      g_EventChatterRealPlayerDistance;
+
+// --------------------------------------------
+// Bot/Player Chatter Probability & Limits
+// --------------------------------------------
+extern uint32_t   g_PlayerReplyChance;
+extern uint32_t   g_BotReplyChance;
+extern uint32_t   g_MaxBotsToPick;
+extern uint32_t   g_RandomChatterBotCommentChance;
+extern uint32_t   g_RandomChatterMaxBotsPerPlayer;
+extern uint32_t   g_EventChatterBotCommentChance;
+extern uint32_t   g_EventChatterBotSelfCommentChance;
+extern uint32_t   g_EventChatterMaxBotsPerPlayer;
+
+// --------------------------------------------
+// LMStudio LLM API Configuration
+// --------------------------------------------
+extern std::string g_LMStudioUrl;
+extern std::string g_LMStudioModel;
+extern uint32_t    g_LMStudioMaxTokens;
+extern float       g_LMStudioTemperature;
+extern float       g_LMStudioTopP;
+extern float       g_LMStudioRepeatPenalty;
+extern uint32_t    g_LMStudioContextSize;
+extern uint32_t    g_LMStudioNumThreads;
+extern std::string g_LMStudioStopSequence;
+extern std::string g_LMStudioSystemPrompt;
+extern std::string g_LMStudioSeed;
+
+// --------------------------------------------
+// Concurrency/Queueing
+// --------------------------------------------
+extern uint32_t    g_MaxConcurrentQueries;
+
+// --------------------------------------------
+// Feature Toggles & Core Settings
+// --------------------------------------------
+extern bool        g_Enable;
+extern bool        g_DisableRepliesInCombat;
+extern bool        g_EnableRandomChatter;
+extern bool        g_EnableEventChatter;
+extern bool        g_EnableRPPersonalities;
+extern bool        g_DebugEnabled;
+
+// --------------------------------------------
+// Random Chatter Timing
+// --------------------------------------------
+extern uint32_t    g_MinRandomInterval;
+extern uint32_t    g_MaxRandomInterval;
+
+// --------------------------------------------
+// Conversation History Settings
+// --------------------------------------------
+extern uint32_t    g_MaxConversationHistory;
+extern uint32_t    g_ConversationHistorySaveInterval;
+
+// --------------------------------------------
+// Prompt Templates
+// --------------------------------------------
+extern std::string g_RandomChatterPromptTemplate;
+extern std::string g_EventChatterPromptTemplate;
+extern std::string g_ChatPromptTemplate;
+extern std::string g_ChatExtraInfoTemplate;
+
+// --------------------------------------------
+// Personality and Prompt Data
+// --------------------------------------------
+extern std::unordered_map<uint64_t, std::string> g_BotPersonalityList;
+extern std::unordered_map<std::string, std::string> g_PersonalityPrompts;
+extern std::vector<std::string> g_PersonalityKeys;
+extern std::string g_DefaultPersonalityPrompt;
+
+// --------------------------------------------
+// Chat History Templates and Toggles
+// --------------------------------------------
+extern bool        g_EnableChatHistory;
+extern std::string g_ChatHistoryHeaderTemplate;
+extern std::string g_ChatHistoryLineTemplate;
+extern std::string g_ChatHistoryFooterTemplate;
+
+// --------------------------------------------
+// Chatbot Snapshot Template
+// --------------------------------------------
+extern bool        g_EnableChatBotSnapshotTemplate;
+extern std::string g_ChatBotSnapshotTemplate;
+
+// --------------------------------------------
+// Conversation History Store and Mutex
+// --------------------------------------------
+extern std::unordered_map<uint64_t, std::unordered_map<uint64_t, std::deque<std::pair<std::string, std::string>>>> g_BotConversationHistory;
+extern std::mutex   g_ConversationHistoryMutex;
+extern time_t       g_LastHistorySaveTime;
+
+// --------------------------------------------
+// Blacklist: Prefixes for Commands (not chat)
+// --------------------------------------------
+extern std::vector<std::string> g_BlacklistCommands;
+
+// --------------------------------------------
+// Think Mode Support
+// --------------------------------------------
+extern bool g_ThinkModeEnableForModule;
+
+// --------------------------------------------
+// Environment/Contextual Random Chatter Templates
+// --------------------------------------------
+extern std::vector<std::string> g_EnvCommentCreature;
+extern std::vector<std::string> g_EnvCommentGameObject;
+extern std::vector<std::string> g_EnvCommentEquippedItem;
+extern std::vector<std::string> g_EnvCommentBagItem;
+extern std::vector<std::string> g_EnvCommentBagItemSell;
+extern std::vector<std::string> g_EnvCommentSpell;
+extern std::vector<std::string> g_EnvCommentQuestArea;
+extern std::vector<std::string> g_EnvCommentVendor;
+extern std::vector<std::string> g_EnvCommentQuestgiver;
+extern std::vector<std::string> g_EnvCommentBagSlots;
+extern std::vector<std::string> g_EnvCommentDungeon;
+extern std::vector<std::string> g_EnvCommentUnfinishedQuest;
+
+// --------------------------------------------
+// Guild-Specific Random Chatter Templates
+// --------------------------------------------
+extern std::vector<std::string> g_GuildEnvCommentGuildMember;
+extern std::vector<std::string> g_GuildEnvCommentGuildRank;
+extern std::vector<std::string> g_GuildEnvCommentGuildBank;
+extern std::vector<std::string> g_GuildEnvCommentGuildMOTD;
+extern std::vector<std::string> g_GuildEnvCommentGuildInfo;
+extern std::vector<std::string> g_GuildEnvCommentGuildOnlineMembers;
+
+// --------------------------------------------
+// Guild-Specific Random Chatter Configuration
+// --------------------------------------------
+extern bool        g_EnableGuildRandomChatter;
+extern uint32_t    g_GuildChatterBotCommentChance;
+extern uint32_t    g_GuildChatterMaxBotsPerEvent;
+
+// --------------------------------------------
+// Guild-Specific Event Chatter Templates
+// --------------------------------------------
+extern std::string g_GuildEventTypeLevelUp;
+extern std::string g_GuildEventTypeDungeonComplete;
+extern std::string g_GuildEventTypeEpicGear;
+extern std::string g_GuildEventTypeRareGear;
+extern std::string g_GuildEventTypeGuildJoin;
+extern std::string g_GuildEventTypeGuildLeave;
+extern std::string g_GuildEventTypeGuildPromotion;
+extern std::string g_GuildEventTypeGuildDemotion;
+
+// --------------------------------------------
+// Bot-Player Sentiment Tracking System
+// --------------------------------------------
+extern bool        g_EnableSentimentTracking;
+extern float       g_SentimentDefaultValue;              // Default sentiment value (0.5 = neutral)
+extern float       g_SentimentAdjustmentStrength;        // How much to adjust sentiment per message (0.1)
+extern uint32_t    g_SentimentSaveInterval;              // How often to save sentiment to DB (minutes)
+extern std::string g_SentimentAnalysisPrompt;            // Prompt template for sentiment analysis
+extern std::string g_SentimentPromptTemplate;            // Template for including sentiment in bot prompts
+
+// In-memory sentiment storage and mutex
+extern std::unordered_map<uint64_t, std::unordered_map<uint64_t, float>> g_BotPlayerSentiments;
+extern std::mutex g_SentimentMutex;
+extern time_t g_LastSentimentSaveTime;
+
+// --------------------------------------------
+// Event Chatter: Event Type Strings
+// These control the event type string sent to eventChatter for world event prompts.
+// Values are loaded from conf (see LMStudioChat.conf.dist)
+// --------------------------------------------
+extern std::string g_EventTypeDefeated;           // "defeated"
+extern std::string g_EventTypeDefeatedPlayer;     // "defeated player"
+extern std::string g_EventTypePetDefeated;        // "pet defeated"
+extern std::string g_EventTypeGotItem;            // "got item"
+extern std::string g_EventTypeDied;               // "died"
+extern std::string g_EventTypeCompletedQuest;     // "completed quest"
+extern std::string g_EventTypeLearnedSpell;       // "learned spell"
+extern std::string g_EventTypeRequestedDuel;      // "requested to duel"
+extern std::string g_EventTypeStartedDueling;     // "started dueling"
+extern std::string g_EventTypeWonDuel;            // "won duel against"
+extern std::string g_EventTypeLeveledUp;          // "leveled up"
+extern std::string g_EventTypeAchievement;        // "earned achievement"
+extern std::string g_EventTypeUsedObject;         // "used object"
+
+// --------------------------------------------
+// Loader Functions
+// --------------------------------------------
+void LoadLMStudioChatConfig();
+void LoadBotPersonalityList();
+void LoadBotConversationHistoryFromDB();
+void LoadPersonalityTemplatesFromDB();
+
+// --------------------------------------------
+// Declaration of the configuration WorldScript.
+// --------------------------------------------
+class LMStudioChatConfigWorldScript : public WorldScript
+{
+public:
+    LMStudioChatConfigWorldScript();
+    void OnStartup() override;
+};
+
+#endif // LMSTUDIO_CHAT_CONFIG_H
+
