+++ src/LMStudioChat_querymanager.h
@@ -1,0 +1,49 @@
+
+
+
+
+
+
+
+
+
+#ifndef LMSTUDIO_CHAT_QUERYMANAGER_H
+#define LMSTUDIO_CHAT_QUERYMANAGER_H
+
+#include <string>
+#include <future>
+#include <mutex>
+#include <queue>
+#include <thread>
+
+std::string QueryLMStudioAPI(const std::string& prompt);
+
+class QueryManager {
+public:
+    QueryManager();
+    void setMaxConcurrentQueries(int maxQueries);
+    std::future<std::string> submitQuery(const std::string& prompt);
+
+private:
+    struct QueryTask {
+        std::string prompt;
+        std::promise<std::string> promise;
+    };
+
+    void processQuery(const std::string& prompt, std::promise<std::string> promise);
+
+    int maxConcurrentQueries; // 0 means no limit
+    int currentQueries;
+    std::mutex mutex_;
+    std::queue<QueryTask> taskQueue;
+};
+
+#endif // LMSTUDIO_CHAT_QUERYMANAGER_H
+
+
+
+
+
+
+
+
