--- src/mod-ollama-chat_config.cpp
@@ -1,525 +1,0 @@
-#include "mod-ollama-chat_config.h"
-#include "mod-ollama-chat_sentiment.h"
-#include "Config.h"
-#include "Log.h"
-#include "mod-ollama-chat_api.h"
-#include <fmt/core.h>
-#include <sstream>
-#include <fstream>
-
-
-// --------------------------------------------
-// Distance/Range Configuration
-// --------------------------------------------
-float      g_SayDistance       = 30.0f;
-float      g_YellDistance      = 100.0f;
-float      g_RandomChatterRealPlayerDistance = 40.0f;
-float      g_EventChatterRealPlayerDistance = 40.0f;
-
-// --------------------------------------------
-// Bot/Player Chatter Probability & Limits
-// --------------------------------------------
-uint32_t   g_PlayerReplyChance = 90;
-uint32_t   g_BotReplyChance    = 10;
-uint32_t   g_MaxBotsToPick     = 2;
-uint32_t   g_RandomChatterBotCommentChance   = 5;
-uint32_t   g_RandomChatterMaxBotsPerPlayer   = 2;
-uint32_t   g_EventChatterBotCommentChance    = 15;
-uint32_t   g_EventChatterBotSelfCommentChance = 5;
-uint32_t   g_EventChatterMaxBotsPerPlayer    = 2;
-
-// --------------------------------------------
-// Ollama LLM API Configuration
-// --------------------------------------------
-std::string g_OllamaUrl        = "http://localhost:11434/api/generate";
-std::string g_OllamaModel      = "llama3.2:1b";
-uint32_t    g_OllamaNumPredict = 40;
-float       g_OllamaTemperature = 0.8f;
-float       g_OllamaTopP = 0.95f;
-float       g_OllamaRepeatPenalty = 1.1f;
-uint32_t    g_OllamaNumCtx = 0;
-uint32_t    g_OllamaNumThreads = 0;
-std::string g_OllamaStop = "";
-std::string g_OllamaSystemPrompt = "";
-std::string g_OllamaSeed = "";
-
-// --------------------------------------------
-// Concurrency/Queueing
-// --------------------------------------------
-uint32_t    g_MaxConcurrentQueries = 0;
-
-// --------------------------------------------
-// Feature Toggles & Core Settings
-// --------------------------------------------
-bool        g_Enable                          = true;
-bool        g_DisableRepliesInCombat          = true;
-bool        g_EnableRandomChatter             = true;
-bool        g_EnableEventChatter              = true;
-bool        g_EnableRPPersonalities           = false;
-bool        g_DebugEnabled                    = false;
-
-// --------------------------------------------
-// Think Mode Support
-// --------------------------------------------
-bool g_ThinkModeEnableForModule = false;
-
-// --------------------------------------------
-// Random Chatter Timing
-// --------------------------------------------
-uint32_t    g_MinRandomInterval               = 45;
-uint32_t    g_MaxRandomInterval               = 180;
-
-// --------------------------------------------
-// Conversation History Settings
-// --------------------------------------------
-uint32_t    g_MaxConversationHistory          = 5;
-uint32_t    g_ConversationHistorySaveInterval = 10;
-
-// --------------------------------------------
-// Prompt Templates
-// --------------------------------------------
-std::string g_RandomChatterPromptTemplate;
-std::string g_EventChatterPromptTemplate;
-std::string g_ChatPromptTemplate;
-std::string g_ChatExtraInfoTemplate;
-
-// --------------------------------------------
-// Personality and Prompt Data
-// --------------------------------------------
-std::unordered_map<uint64_t, std::string> g_BotPersonalityList;
-std::unordered_map<std::string, std::string> g_PersonalityPrompts;
-std::vector<std::string> g_PersonalityKeys;
-std::string g_DefaultPersonalityPrompt;
-
-// --------------------------------------------
-// Chat History Templates and Toggles
-// --------------------------------------------
-bool        g_EnableChatHistory = true;
-std::string g_ChatHistoryHeaderTemplate;
-std::string g_ChatHistoryLineTemplate;
-std::string g_ChatHistoryFooterTemplate;
-
-// --------------------------------------------
-// Chatbot Snapshot Template
-// --------------------------------------------
-bool        g_EnableChatBotSnapshotTemplate  = false;
-std::string g_ChatBotSnapshotTemplate;
-
-// --------------------------------------------
-// Conversation History Store and Mutex
-// --------------------------------------------
-std::unordered_map<uint64_t, std::unordered_map<uint64_t, std::deque<std::pair<std::string, std::string>>>> g_BotConversationHistory;
-std::mutex g_ConversationHistoryMutex;
-time_t g_LastHistorySaveTime = 0;
-
-// --------------------------------------------
-// Bot-Player Sentiment Tracking System
-// --------------------------------------------
-bool        g_EnableSentimentTracking = true;
-float       g_SentimentDefaultValue = 0.5f;              // Default sentiment value (0.5 = neutral)
-float       g_SentimentAdjustmentStrength = 0.1f;        // How much to adjust sentiment per message
-uint32_t    g_SentimentSaveInterval = 10;                // How often to save sentiment to DB (minutes)
-std::string g_SentimentAnalysisPrompt = "Analyze the sentiment of this message: \"{message}\". Respond only with: POSITIVE, NEGATIVE, or NEUTRAL.";
-std::string g_SentimentPromptTemplate = "Your relationship sentiment with {player_name} is {sentiment_value} (0.0=hostile, 0.5=neutral, 1.0=friendly). Use this to guide your tone and response.";
-
-// In-memory sentiment storage and mutex
-std::unordered_map<uint64_t, std::unordered_map<uint64_t, float>> g_BotPlayerSentiments;
-std::mutex g_SentimentMutex;
-time_t g_LastSentimentSaveTime = 0;
-
-// --------------------------------------------
-// Blacklist: Prefixes for Commands (not chat)
-// --------------------------------------------
-std::vector<std::string> g_BlacklistCommands = {
-    ".playerbots",
-    "playerbot",
-};
-
-// --------------------------------------------
-// Environment/Contextual Random Chatter Templates
-// --------------------------------------------
-std::vector<std::string> g_EnvCommentCreature;
-std::vector<std::string> g_EnvCommentGameObject;
-std::vector<std::string> g_EnvCommentEquippedItem;
-std::vector<std::string> g_EnvCommentBagItem;
-std::vector<std::string> g_EnvCommentBagItemSell;
-std::vector<std::string> g_EnvCommentSpell;
-std::vector<std::string> g_EnvCommentQuestArea;
-std::vector<std::string> g_EnvCommentVendor;
-std::vector<std::string> g_EnvCommentQuestgiver;
-std::vector<std::string> g_EnvCommentBagSlots;
-std::vector<std::string> g_EnvCommentDungeon;
-std::vector<std::string> g_EnvCommentUnfinishedQuest;
-
-// --------------------------------------------
-// Guild-Specific Random Chatter Templates
-// --------------------------------------------
-std::vector<std::string> g_GuildEnvCommentGuildMember;
-std::vector<std::string> g_GuildEnvCommentGuildRank;
-std::vector<std::string> g_GuildEnvCommentGuildBank;
-std::vector<std::string> g_GuildEnvCommentGuildMOTD;
-std::vector<std::string> g_GuildEnvCommentGuildInfo;
-std::vector<std::string> g_GuildEnvCommentGuildOnlineMembers;
-
-// --------------------------------------------
-// Guild-Specific Random Chatter Configuration
-// --------------------------------------------
-bool        g_EnableGuildRandomChatter             = true;
-uint32_t    g_GuildChatterBotCommentChance          = 25;
-uint32_t    g_GuildChatterMaxBotsPerEvent           = 2;
-
-// --------------------------------------------
-// Guild-Specific Event Chatter Templates
-// --------------------------------------------
-std::string g_GuildEventTypeLevelUp = "";
-std::string g_GuildEventTypeDungeonComplete = "";
-std::string g_GuildEventTypeEpicGear = "";
-std::string g_GuildEventTypeRareGear = "";
-std::string g_GuildEventTypeGuildJoin = "";
-std::string g_GuildEventTypeGuildLeave = "";
-std::string g_GuildEventTypeGuildPromotion = "";
-std::string g_GuildEventTypeGuildDemotion = "";
-
-// --------------------------------------------
-// Event Chatter Templates
-// --------------------------------------------
-std::string g_EventTypeDefeated;           // "defeated"
-std::string g_EventTypeDefeatedPlayer;     // "defeated player"
-std::string g_EventTypePetDefeated;        // "pet defeated"
-std::string g_EventTypeGotItem;            // "got item"
-std::string g_EventTypeDied;               // "died"
-std::string g_EventTypeCompletedQuest;     // "completed quest"
-std::string g_EventTypeLearnedSpell;       // "learned spell"
-std::string g_EventTypeRequestedDuel;      // "requested to duel"
-std::string g_EventTypeStartedDueling;     // "started dueling"
-std::string g_EventTypeWonDuel;            // "won duel against"
-std::string g_EventTypeLeveledUp;          // "leveled up"
-std::string g_EventTypeAchievement;        // "earned achievement"
-std::string g_EventTypeUsedObject;         // "used object"
-
-
-static std::vector<std::string> SplitString(const std::string& str, char delim)
-{
-    std::vector<std::string> tokens;
-    std::stringstream ss(str);
-    std::string token;
-    while (std::getline(ss, token, delim))
-    {
-        // Trim whitespace from token
-        size_t start = token.find_first_not_of(" \t");
-        size_t end = token.find_last_not_of(" \t");
-        if (start != std::string::npos && end != std::string::npos)
-            tokens.push_back(token.substr(start, end - start + 1));
-    }
-    return tokens;
-}
-
-// Load Bot Personalities from Database
-void LoadBotPersonalityList()
-{    
-    // Let's make sure our user has sourced the required sql file to add the new table
-    QueryResult tableExists = CharacterDatabase.Query("SELECT * FROM information_schema.tables WHERE table_schema = 'acore_characters' AND table_name = 'mod_ollama_chat_personality' LIMIT 1");
-    if (!tableExists)
-    {
-        LOG_ERROR("server.loading", "[Ollama Chat] Please source the required database table first");
-        return;
-    }
-
-    QueryResult result = CharacterDatabase.Query("SELECT guid,personality FROM mod_ollama_chat_personality");
-
-    if (!result)
-    {
-        return;
-    }
-    if (result->GetRowCount() == 0)
-    {
-        return;
-    }    
-
-    if(g_DebugEnabled)
-    {
-        LOG_INFO("server.loading", "[Ollama Chat] Fetching Bot Personality List into array");
-    }
-
-    do
-    {
-        uint64_t personalityBotGUID = result->Fetch()[0].Get<uint64_t>();
-        std::string personalityKey = result->Fetch()[1].Get<std::string>();
-        g_BotPersonalityList[personalityBotGUID] = personalityKey;
-    } while (result->NextRow());
-}
-
-std::string GetMultiLineConfigValue(const std::string& configFilePath, const std::string& key)
-{
-    std::ifstream infile(configFilePath);
-    if (!infile) return "";
-
-    std::string line;
-    std::string value;
-    bool foundKey = false;
-    while (std::getline(infile, line))
-    {
-        std::string trimmed = line;
-        trimmed.erase(0, trimmed.find_first_not_of(" \t\r\n"));
-        if (trimmed.empty() || trimmed[0] == '#')
-            continue;
-        size_t pos = trimmed.find('=');
-        if (!foundKey && pos != std::string::npos) {
-            std::string possibleKey = trimmed.substr(0, pos);
-            possibleKey.erase(possibleKey.find_last_not_of(" \t\r\n") + 1);
-            if (possibleKey == key) {
-                foundKey = true;
-                std::string afterEq = trimmed.substr(pos + 1);
-                afterEq.erase(0, afterEq.find_first_not_of(" \t\r\n"));
-                value += afterEq;
-                continue;
-            }
-        }
-        else if (foundKey) {
-            // New config key or section
-            if (trimmed.find('=') != std::string::npos && trimmed.find('[') == std::string::npos)
-                break;
-            if (!value.empty()) value += "\n";
-            value += trimmed;
-        }
-    }
-
-    return value;
-}
-
-void LoadOllamaChatConfig()
-{
-    g_SayDistance                     = sConfigMgr->GetOption<float>("OllamaChat.SayDistance", 30.0f);
-    g_YellDistance                    = sConfigMgr->GetOption<float>("OllamaChat.YellDistance", 100.0f);
-    g_PlayerReplyChance               = sConfigMgr->GetOption<uint32_t>("OllamaChat.PlayerReplyChance", 90);
-    g_BotReplyChance                  = sConfigMgr->GetOption<uint32_t>("OllamaChat.BotReplyChance", 10);
-    g_MaxBotsToPick                   = sConfigMgr->GetOption<uint32_t>("OllamaChat.MaxBotsToPick", 2);
-    g_OllamaUrl                       = sConfigMgr->GetOption<std::string>("OllamaChat.Url", "http://localhost:11434/api/generate");
-    g_OllamaModel                     = sConfigMgr->GetOption<std::string>("OllamaChat.Model", "llama3.2:1b");
-    g_OllamaNumPredict                = sConfigMgr->GetOption<uint32_t>("OllamaChat.NumPredict", 40);
-    g_OllamaTemperature               = sConfigMgr->GetOption<float>("OllamaChat.Temperature", 0.8f);
-    g_OllamaTopP                      = sConfigMgr->GetOption<float>("OllamaChat.TopP", 0.95f);
-    g_OllamaRepeatPenalty             = sConfigMgr->GetOption<float>("OllamaChat.RepeatPenalty", 1.1f);
-    g_OllamaNumCtx                    = sConfigMgr->GetOption<uint32_t>("OllamaChat.NumCtx", 0);
-    g_OllamaNumThreads                = sConfigMgr->GetOption<uint32_t>("OllamaChat.NumThreads", 0);
-    g_OllamaStop                      = sConfigMgr->GetOption<std::string>("OllamaChat.Stop", "");
-    g_OllamaSystemPrompt              = sConfigMgr->GetOption<std::string>("OllamaChat.SystemPrompt", "");
-    g_OllamaSeed                      = sConfigMgr->GetOption<std::string>("OllamaChat.Seed", "");
-
-    g_MaxConcurrentQueries            = sConfigMgr->GetOption<uint32_t>("OllamaChat.MaxConcurrentQueries", 0);
-
-    g_Enable                          = sConfigMgr->GetOption<bool>("OllamaChat.Enable", true);
-    g_DisableRepliesInCombat          = sConfigMgr->GetOption<bool>("OllamaChat.DisableRepliesInCombat", true);
-    g_EnableRandomChatter             = sConfigMgr->GetOption<bool>("OllamaChat.EnableRandomChatter", true);
-    g_EnableEventChatter              = sConfigMgr->GetOption<bool>("OllamaChat.EnableEventChatter", true);
-
-    g_DebugEnabled                    = sConfigMgr->GetOption<bool>("OllamaChat.DebugEnabled", false);
-
-    g_MinRandomInterval               = sConfigMgr->GetOption<uint32_t>("OllamaChat.MinRandomInterval", 45);
-    g_MaxRandomInterval               = sConfigMgr->GetOption<uint32_t>("OllamaChat.MaxRandomInterval", 180);
-    g_RandomChatterRealPlayerDistance = sConfigMgr->GetOption<float>("OllamaChat.RandomChatterRealPlayerDistance", 40.0f);
-    g_RandomChatterBotCommentChance   = sConfigMgr->GetOption<uint32_t>("OllamaChat.RandomChatterBotCommentChance", 25);
-    g_RandomChatterMaxBotsPerPlayer   = sConfigMgr->GetOption<uint32_t>("OllamaChat.RandomChatterMaxBotsPerPlayer", 2);
-
-    g_EventChatterRealPlayerDistance = sConfigMgr->GetOption<float>("OllamaChat.EventChatterRealPlayerDistance", 40.0f);
-    g_EventChatterBotCommentChance   = sConfigMgr->GetOption<uint32_t>("OllamaChat.EventChatterBotCommentChance", 15);
-    g_EventChatterBotSelfCommentChance = sConfigMgr->GetOption<uint32_t>("OllamaChat.EventChatterBotSelfCommentChance", 5);
-    g_EventChatterMaxBotsPerPlayer   = sConfigMgr->GetOption<uint32_t>("OllamaChat.EventChatterMaxBotsPerPlayer", 2);
-
-    g_EnableRPPersonalities           = sConfigMgr->GetOption<bool>("OllamaChat.EnableRPPersonalities", false);
-
-    g_RandomChatterPromptTemplate     = sConfigMgr->GetOption<std::string>("OllamaChat.RandomChatterPromptTemplate", "");
-
-    g_EventChatterPromptTemplate     = sConfigMgr->GetOption<std::string>("OllamaChat.EventChatterPromptTemplate", "");
-
-    g_ChatPromptTemplate              = sConfigMgr->GetOption<std::string>("OllamaChat.ChatPromptTemplate", "");
-    
-    g_ChatExtraInfoTemplate           = sConfigMgr->GetOption<std::string>("OllamaChat.ChatExtraInfoTemplate", "");
-
-    g_DefaultPersonalityPrompt        = sConfigMgr->GetOption<std::string>("OllamaChat.DefaultPersonalityPrompt", "");
-
-    g_MaxConversationHistory          = sConfigMgr->GetOption<uint32_t>("OllamaChat.MaxConversationHistory", 5);
-    g_ConversationHistorySaveInterval = sConfigMgr->GetOption<uint32_t>("OllamaChat.ConversationHistorySaveInterval", 10);
-
-    g_ChatHistoryHeaderTemplate       = sConfigMgr->GetOption<std::string>("OllamaChat.ChatHistoryHeaderTemplate", "");
-    g_ChatHistoryLineTemplate         = sConfigMgr->GetOption<std::string>("OllamaChat.ChatHistoryLineTemplate", "");
-    g_ChatHistoryFooterTemplate       = sConfigMgr->GetOption<std::string>("OllamaChat.ChatHistoryFooterTemplate", "");
-
-    g_EnableChatBotSnapshotTemplate   = sConfigMgr->GetOption<bool>("OllamaChat.EnableChatBotSnapshotTemplate", false);
-    g_ChatBotSnapshotTemplate         = sConfigMgr->GetOption<std::string>("OllamaChat.ChatBotSnapshotTemplate", "");
-
-    g_EnableChatHistory               = sConfigMgr->GetOption<bool>("OllamaChat.EnableChatHistory", true);
-
-    // Bot-Player Sentiment Tracking
-    g_EnableSentimentTracking         = sConfigMgr->GetOption<bool>("OllamaChat.EnableSentimentTracking", true);
-    g_SentimentDefaultValue           = sConfigMgr->GetOption<float>("OllamaChat.SentimentDefaultValue", 0.5f);
-    g_SentimentAdjustmentStrength     = sConfigMgr->GetOption<float>("OllamaChat.SentimentAdjustmentStrength", 0.1f);
-    g_SentimentSaveInterval           = sConfigMgr->GetOption<uint32_t>("OllamaChat.SentimentSaveInterval", 10);
-    g_SentimentAnalysisPrompt         = sConfigMgr->GetOption<std::string>("OllamaChat.SentimentAnalysisPrompt", "Analyze the sentiment of this message: \"{message}\". Respond only with: POSITIVE, NEGATIVE, or NEUTRAL.");
-    g_SentimentPromptTemplate         = sConfigMgr->GetOption<std::string>("OllamaChat.SentimentPromptTemplate", "Your relationship sentiment with {player_name} is {sentiment_value} (0.0=hostile, 0.5=neutral, 1.0=friendly). Use this to guide your tone and response.");
-
-    g_ThinkModeEnableForModule        = sConfigMgr->GetOption<bool>("OllamaChat.ThinkModeEnableForModule", false);
-
-    g_EventTypeDefeated           = sConfigMgr->GetOption<std::string>("OllamaChat.EventTypeDefeated", "");
-    g_EventTypeDefeatedPlayer     = sConfigMgr->GetOption<std::string>("OllamaChat.EventTypeDefeatedPlayer", "");
-    g_EventTypePetDefeated        = sConfigMgr->GetOption<std::string>("OllamaChat.EventTypePetDefeated", "");
-    g_EventTypeGotItem            = sConfigMgr->GetOption<std::string>("OllamaChat.EventTypeGotItem", "");
-    g_EventTypeDied               = sConfigMgr->GetOption<std::string>("OllamaChat.EventTypeDied", "");
-    g_EventTypeCompletedQuest     = sConfigMgr->GetOption<std::string>("OllamaChat.EventTypeCompletedQuest", "");
-    g_EventTypeLearnedSpell       = sConfigMgr->GetOption<std::string>("OllamaChat.EventTypeLearnedSpell", "");
-    g_EventTypeRequestedDuel      = sConfigMgr->GetOption<std::string>("OllamaChat.EventTypeRequestedDuel", "");
-    g_EventTypeStartedDueling     = sConfigMgr->GetOption<std::string>("OllamaChat.EventTypeStartedDueling", "");
-    g_EventTypeWonDuel            = sConfigMgr->GetOption<std::string>("OllamaChat.EventTypeWonDuel", "");
-    g_EventTypeLeveledUp          = sConfigMgr->GetOption<std::string>("OllamaChat.EventTypeLeveledUp", "");
-    g_EventTypeAchievement        = sConfigMgr->GetOption<std::string>("OllamaChat.EventTypeAchievement", "");
-    g_EventTypeUsedObject         = sConfigMgr->GetOption<std::string>("OllamaChat.EventTypeUsedObject", "");
-
-
-    // Load extra blacklist commands from config (comma-separated list)
-    std::string extraBlacklist = sConfigMgr->GetOption<std::string>("OllamaChat.BlacklistCommands", "");
-    if (!extraBlacklist.empty())
-    {
-        std::vector<std::string> extraList = SplitString(extraBlacklist, ',');
-        for (const auto& cmd : extraList)
-        {
-            g_BlacklistCommands.push_back(cmd);
-        }
-    }
-
-    LoadPersonalityTemplatesFromDB();
-
-    g_queryManager.setMaxConcurrentQueries(g_MaxConcurrentQueries);
-
-    // Loads the environment random chatter message templates for each type.
-    // Each config option is a pipe-separated list of string templates,
-    // using {} as a placeholder for named substitutions.
-    // Helper to load a multi-line config option into a std::vector<std::string>
-    auto LoadEnvCommentVector = [](const char* key, const std::vector<std::string>& defaults = {}) -> std::vector<std::string>
-    {
-        std::string val = sConfigMgr->GetOption<std::string>(key, "");
-        std::vector<std::string> result;
-        std::istringstream iss(val);
-        std::string line;
-        while (std::getline(iss, line)) {
-            // Trim whitespace (both ends)
-            size_t start = line.find_first_not_of(" \t\r\n");
-            size_t end = line.find_last_not_of(" \t\r\n");
-            if (start != std::string::npos && end != std::string::npos)
-                result.push_back(line.substr(start, end - start + 1));
-        }
-        if (result.empty() && !defaults.empty())
-            return defaults;
-        return result;
-    };
-
-    g_EnvCommentCreature        = LoadEnvCommentVector("OllamaChat.EnvCommentCreature", { "" });
-    g_EnvCommentGameObject      = LoadEnvCommentVector("OllamaChat.EnvCommentGameObject", { "" });
-    g_EnvCommentEquippedItem    = LoadEnvCommentVector("OllamaChat.EnvCommentEquippedItem", { "" });
-    g_EnvCommentBagItem         = LoadEnvCommentVector("OllamaChat.EnvCommentBagItem", { "" });
-    g_EnvCommentBagItemSell     = LoadEnvCommentVector("OllamaChat.EnvCommentBagItemSell", { "" });
-    g_EnvCommentSpell           = LoadEnvCommentVector("OllamaChat.EnvCommentSpell", { "" });
-    g_EnvCommentQuestArea       = LoadEnvCommentVector("OllamaChat.EnvCommentQuestArea", { "" });
-    g_EnvCommentVendor          = LoadEnvCommentVector("OllamaChat.EnvCommentVendor", { "" });
-    g_EnvCommentQuestgiver      = LoadEnvCommentVector("OllamaChat.EnvCommentQuestgiver", { "" });
-    g_EnvCommentBagSlots        = LoadEnvCommentVector("OllamaChat.EnvCommentBagSlots", { "" });
-    g_EnvCommentDungeon         = LoadEnvCommentVector("OllamaChat.EnvCommentDungeon", { "" });
-    g_EnvCommentUnfinishedQuest = LoadEnvCommentVector("OllamaChat.EnvCommentUnfinishedQuest", { "" });
-
-    // Guild-specific random chatter templates
-    g_GuildEnvCommentGuildMember = LoadEnvCommentVector("OllamaChat.GuildEnvCommentGuildMember", { "" });
-    g_GuildEnvCommentGuildRank = LoadEnvCommentVector("OllamaChat.GuildEnvCommentGuildRank", { "" });
-    g_GuildEnvCommentGuildBank = LoadEnvCommentVector("OllamaChat.GuildEnvCommentGuildBank", { "" });
-    g_GuildEnvCommentGuildMOTD = LoadEnvCommentVector("OllamaChat.GuildEnvCommentGuildMOTD", { "" });
-    g_GuildEnvCommentGuildInfo = LoadEnvCommentVector("OllamaChat.GuildEnvCommentGuildInfo", { "" });
-    g_GuildEnvCommentGuildOnlineMembers = LoadEnvCommentVector("OllamaChat.GuildEnvCommentGuildOnlineMembers", { "" });
-
-    // Guild-specific configuration
-    g_EnableGuildRandomChatter = sConfigMgr->GetOption<bool>("OllamaChat.EnableGuildRandomChatter", true);
-    g_GuildChatterBotCommentChance = sConfigMgr->GetOption<uint32_t>("OllamaChat.GuildChatterBotCommentChance", 25);
-    g_GuildChatterMaxBotsPerEvent = sConfigMgr->GetOption<uint32_t>("OllamaChat.GuildChatterMaxBotsPerEvent", 2);
-
-    // Guild-specific event templates
-    g_GuildEventTypeLevelUp = sConfigMgr->GetOption<std::string>("OllamaChat.GuildEventTypeLevelUp", "");
-    g_GuildEventTypeDungeonComplete = sConfigMgr->GetOption<std::string>("OllamaChat.GuildEventTypeDungeonComplete", "");
-    g_GuildEventTypeEpicGear = sConfigMgr->GetOption<std::string>("OllamaChat.GuildEventTypeEpicGear", "");
-    g_GuildEventTypeRareGear = sConfigMgr->GetOption<std::string>("OllamaChat.GuildEventTypeRareGear", "");
-    g_GuildEventTypeGuildJoin = sConfigMgr->GetOption<std::string>("OllamaChat.GuildEventTypeGuildJoin", "");
-    g_GuildEventTypeGuildLeave = sConfigMgr->GetOption<std::string>("OllamaChat.GuildEventTypeGuildLeave", "");
-    g_GuildEventTypeGuildPromotion = sConfigMgr->GetOption<std::string>("OllamaChat.GuildEventTypeGuildPromotion", "");
-    g_GuildEventTypeGuildDemotion = sConfigMgr->GetOption<std::string>("OllamaChat.GuildEventTypeGuildDemotion", "");
-
-    LOG_INFO("server.loading",
-             "[Ollama Chat] Config loaded: Enabled = {}, SayDistance = {}, YellDistance = {}, "
-             "PlayerReplyChance = {}%, BotReplyChance = {}%, MaxBotsToPick = {}, "
-             "Url = {}, Model = {}, MaxConcurrentQueries = {}, EnableRandomChatter = {}, MinRandInt = {}, MaxRandInt = {}, RandomChatterRealPlayerDistance = {}, "
-             "RandomChatterBotCommentChance = {}. MaxConcurrentQueries = {}. Extra blacklist commands: {}",
-             g_Enable, g_SayDistance, g_YellDistance,
-             g_PlayerReplyChance, g_BotReplyChance, g_MaxBotsToPick,
-             g_OllamaUrl, g_OllamaModel, g_MaxConcurrentQueries,
-             g_EnableRandomChatter, g_MinRandomInterval, g_MaxRandomInterval, g_RandomChatterRealPlayerDistance,
-             g_RandomChatterBotCommentChance, g_MaxConcurrentQueries, extraBlacklist);
-}
-
-void LoadPersonalityTemplatesFromDB()
-{
-    g_PersonalityPrompts.clear();
-    g_PersonalityKeys.clear();
-
-    QueryResult result = CharacterDatabase.Query("SELECT `key`, `prompt` FROM `mod_ollama_chat_personality_templates`");
-    if (!result)
-    {
-        LOG_ERROR("server.loading", "[Ollama Chat] No personality templates found in the database!");
-        return;
-    }
-
-    do
-    {
-        std::string key = (*result)[0].Get<std::string>();
-        std::string prompt = (*result)[1].Get<std::string>();
-        g_PersonalityPrompts[key] = prompt;
-        g_PersonalityKeys.push_back(key);
-    } while (result->NextRow());
-
-    LOG_INFO("server.loading", "[Ollama Chat] Cached {} personalities.", g_PersonalityKeys.size());
-
-}
-
-void LoadBotConversationHistoryFromDB()
-{
-    QueryResult result = CharacterDatabase.Query(
-        "SELECT bot_guid, player_guid, player_message, bot_reply FROM mod_ollama_chat_history ORDER BY timestamp ASC"
-    );
-    if (!result)
-        return;
-
-    std::lock_guard<std::mutex> lock(g_ConversationHistoryMutex);
-    g_BotConversationHistory.clear();
-
-    do {
-        uint64_t botGuid = (*result)[0].Get<uint64_t>();
-        uint64_t playerGuid = (*result)[1].Get<uint64_t>();
-        std::string playerMsg = (*result)[2].Get<std::string>();
-        std::string botReply = (*result)[3].Get<std::string>();
-
-        auto& playerHistory = g_BotConversationHistory[botGuid][playerGuid];
-        playerHistory.push_back({ playerMsg, botReply });
-        while (playerHistory.size() > g_MaxConversationHistory)
-        {
-            playerHistory.pop_front();
-        }
-
-    } while (result->NextRow());
-
-}
-
-
-// Definition of the configuration WorldScript.
-OllamaChatConfigWorldScript::OllamaChatConfigWorldScript() : WorldScript("OllamaChatConfigWorldScript") { }
-
-void OllamaChatConfigWorldScript::OnStartup()
-{
-    LoadOllamaChatConfig();
-    LoadBotPersonalityList();
-    LoadBotConversationHistoryFromDB();
-    InitializeSentimentTracking();
-}
